requirements.txt
azure-eventhub==5.0.1
azure.storage.blob
certifi
colorama>=0.4.1
docker>=3.5.0
junitparser==1.4.1
msrest
pre-commit
pytest==5.4.2
pytest-asyncio==0.12.0
pytest-repeat
pytest-testdox
pytest-timeout
python_dateutil
requests>=2.20.0
ruamel.yaml>=0.15.71
setuptools
six
uamqp
urllib3>=1.25.8
PyYAML>=5.3.1
inflection==0.5.0
